rag_gemini_chroma/requirements.txt
fastapi
uvicorn
streamlit
requests
python-dotenv
python-multipart
chromadb
pypdf
tqdm
langchain-text-splitters
sentence-transformers
google-genai
